requirements.txt
openai
PyPDF2
python-dotenv
langchain_community
unstructured
unstructured[all-docs]
openai==0.28